requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-core
google-generativeai
sentence-transformers
faiss-cpu
youtube-transcript-api
pytube
python-dotenv
streamlit
pydantic